+++ aiohttp_sqlalchemy/__init__.py
@@ -39,7 +39,7 @@ from aiohttp_sqlalchemy.views import (
     SAView,
 )
 
-__version__ = '0.20.1'
+__version__ = '0.21.0'
 
 __all__ = [
     'SA_DEFAULT_KEY',